codebase/requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Data science & metrics
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing (for GradCAM overlay)
opencv-python>=4.8.0

# Progress bars
tqdm>=4.65.0

# Image loading
Pillow>=10.0.0
